Реализует открытие меню сортировки, оптимизирует компонент
